arquillian-infrastructure: Refactor
