+++ src/test/resources/webapp/columnIT.xhtml
@@ -11,10 +11,13 @@
     <h:body>
        <b:container>
             <b:row>
-                <b:column id="col_1" col-xs="1" col-sm="0" col-lg="0" ></b:column>
-                <b:column id="col_2" col-xs="0" col-sm="1" col-lg="0"></b:column>
-                <b:column id="col_3" col-xs="0" col-sm="0" col-lg="1"></b:column>
-                <b:column id="col_4"></b:column>
+                <b:column id="col_1" col-xs="0" col-sm="0" col-lg="0" ></b:column>
+                <b:column id="col_2" col-xs="1" col-sm="1" col-lg="1"></b:column>
+                <b:column id="col_3" col-xs="1" col-sm="1" col-lg="0"></b:column>
+                <b:column id="col_4" col-xs="1" col-sm="0" col-lg="1"></b:column>
+                <b:column id="col_5" col-xs="0" col-sm="1" col-lg="1"></b:column>
+                           
+                <b:column id="col_6"></b:column>
             </b:row>
         </b:container>
     </h:body>
